pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-ledger"
version = "0.1.0"
description = "Portfolio-level project tracking for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Rob Leach" },
]
keywords = ["claude-code", "portfolio", "project-tracking", "ledger"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
claude-ledger = "claude_ledger.cli:cli"

[project.urls]
Homepage = "https://github.com/claude-ledger/claude-ledger"
Repository = "https://github.com/claude-ledger/claude-ledger"
Issues = "https://github.com/claude-ledger/claude-ledger/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_ledger"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
